requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-storage-blob
azure-identity
pyodbc
openpyxl
urllib3
SQLAlchemy           
pysqlite3               
scipy
beautifulsoup4
numpy
pandas